refactor: split app control from terminal UI
