+++ CMakeLists.txt
@@ -968,12 +968,13 @@ add_subdirectory(src/core)
 add_subdirectory(src/dsp)
 add_subdirectory(src/io)
 add_subdirectory(src/runtime)
-add_subdirectory(src/ui)
 add_subdirectory(src/protocol)
 add_subdirectory(src/crypto)
 add_subdirectory(src/fec)
 add_subdirectory(src/third_party)
 add_subdirectory(src/engine)
+add_subdirectory(src/app_control)
+add_subdirectory(src/ui)
 add_subdirectory(apps)
 # Unit tests are built when BUILD_TESTING is enabled.
 if(BUILD_TESTING)
