+++ include/dsd-neo/engine/engine.h
@@ -13,7 +13,26 @@
 extern "C" {
 #endif
 
+typedef int (*dsd_engine_lifecycle_start_fn)(dsd_opts* opts, dsd_state* state, void* context);
+typedef void (*dsd_engine_lifecycle_stop_fn)(dsd_opts* opts, dsd_state* state, void* context);
+
+typedef struct {
+    dsd_engine_lifecycle_start_fn start;
+    dsd_engine_lifecycle_stop_fn stop;
+    void* context;
+} dsd_engine_lifecycle_hooks;
+
 int dsd_engine_run(dsd_opts* opts, dsd_state* state);
+
+/**
+ * Run the engine with optional live-run lifecycle hooks.
+ *
+ * The start hook is called after engine and mode-specific setup succeeds, just
+ * before live processing begins. If start succeeds, the stop hook is called
+ * after live processing returns and before dsd_engine_cleanup() tears down
+ * engine-owned state.
+ */
+int dsd_engine_run_with_lifecycle(dsd_opts* opts, dsd_state* state, const dsd_engine_lifecycle_hooks* hooks);
 void dsd_engine_cleanup(dsd_opts* opts, dsd_state* state);
 
 #ifdef __cplusplus
