+++ include/dsd-neo/ui/ui_cmd_dispatch.h
@@ -5,37 +5,12 @@
 
 /**
  * @file
- * @brief UI command dispatch registry and handler signature.
+ * @brief Compatibility wrapper for app-control command dispatch declarations.
  */
 
 #ifndef DSD_NEO_INCLUDE_DSD_NEO_UI_UI_CMD_DISPATCH_H_
 #define DSD_NEO_INCLUDE_DSD_NEO_UI_UI_CMD_DISPATCH_H_
 
-#include <dsd-neo/core/opts_fwd.h>
-#include <dsd-neo/core/state_fwd.h>
-#include <dsd-neo/ui/ui_cmd.h>
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/** Common handler signature for UI command actions. */
-typedef int (*UiCmdHandler)(dsd_opts* opts, dsd_state* state, const struct UiCmd* cmd);
-
-/** Registry entry mapping command id -> handler function. */
-struct UiCmdReg {
-    int id;
-    UiCmdHandler fn;
-};
-
-/** Per-domain registries (terminated with `{0, NULL}`). */
-extern const struct UiCmdReg ui_actions_audio[];
-extern const struct UiCmdReg ui_actions_radio[];
-extern const struct UiCmdReg ui_actions_trunk[];
-extern const struct UiCmdReg ui_actions_logging[];
-
-#ifdef __cplusplus
-}
-#endif
+#include "dsd-neo/app_control/command_dispatch.h" // IWYU pragma: export
 
 #endif /* DSD_NEO_INCLUDE_DSD_NEO_UI_UI_CMD_DISPATCH_H_ */
