+++ include/dsd-neo/ui/ui_opts_snapshot.h
@@ -5,21 +5,18 @@
 
 /**
  * @file
- * @brief Snapshot API for publishing dsd_opts to the UI thread.
+ * @brief Compatibility wrapper for app-control options snapshots.
  */
 #ifndef DSD_NEO_INCLUDE_DSD_NEO_UI_UI_OPTS_SNAPSHOT_H_
 #define DSD_NEO_INCLUDE_DSD_NEO_UI_UI_OPTS_SNAPSHOT_H_
 
-#include <dsd-neo/core/opts_fwd.h>
+#include "dsd-neo/app_control/snapshot.h" // IWYU pragma: export
+#include "dsd-neo/core/opts_fwd.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/**
- * @brief Get the latest published options snapshot.
- * @return Pointer to stable copy, or NULL if none published yet.
- */
 const dsd_opts* ui_get_latest_opts_snapshot(void);
 
 #ifdef __cplusplus
