+++ src/app_control/actions/actions_logging.c
@@ -5,15 +5,15 @@
 
 /* UI command actions — logging/history domain */
 
+#include <dsd-neo/app_control/command_dispatch.h>
+#include <dsd-neo/app_control/services.h>
 #include <dsd-neo/core/state.h>
-#include <dsd-neo/ui/menu_services.h>
-#include <dsd-neo/ui/ui_cmd_dispatch.h>
 #include <string.h>
 #include <time.h>
+#include "dsd-neo/app_control/commands.h"
 #include "dsd-neo/core/opts_fwd.h"
 #include "dsd-neo/core/safe_api.h"
 #include "dsd-neo/core/state_fwd.h"
-#include "dsd-neo/ui/ui_cmd.h"
 
 static int
 ui_handle_eh_next(dsd_opts* opts, dsd_state* state, const struct UiCmd* c) {
