+++ src/app_control/actions/actions_radio.c
@@ -5,16 +5,16 @@
 
 /* UI command actions — radio domain */
 
+#include <dsd-neo/app_control/command_dispatch.h>
 #include <dsd-neo/core/opts.h>
 #include <dsd-neo/core/state.h>
 #include <dsd-neo/io/rtl_stream_c.h>
-#include <dsd-neo/ui/ui_cmd_dispatch.h>
 #include <stdint.h>
 #include <string.h>
+#include "dsd-neo/app_control/commands.h"
 #include "dsd-neo/core/opts_fwd.h"
 #include "dsd-neo/core/safe_api.h"
 #include "dsd-neo/core/state_fwd.h"
-#include "dsd-neo/ui/ui_cmd.h"
 
 #ifdef USE_RADIO
 #endif
