+++ src/ui/terminal/dsd_ncurses_printer.c
@@ -19,6 +19,7 @@
 #include "dsd-neo/core/input_level.h"
 
 #include <curses.h>
+#include <dsd-neo/app_control/history.h>
 #include <dsd-neo/core/opts.h>
 #include <dsd-neo/core/power.h>
 #include <dsd-neo/core/state.h>
@@ -38,7 +39,6 @@
 #include <dsd-neo/ui/ncurses_trunk_display.h>
 #include <dsd-neo/ui/panels.h>
 #include <dsd-neo/ui/ui_async.h>
-#include <dsd-neo/ui/ui_history.h>
 #include <dsd-neo/ui/ui_prims.h>
 #include <stdarg.h>
 #include <stdint.h>
