+++ src/ui/terminal/menu_callbacks.c
@@ -9,12 +9,11 @@
  */
 
 #include "menu_callbacks.h"
+#include <dsd-neo/app_control/commands.h>
 #include <dsd-neo/core/opts.h>
 #include <dsd-neo/core/state.h>
 #include <dsd-neo/platform/posix_compat.h>
 #include <dsd-neo/runtime/config.h>
-#include <dsd-neo/ui/ui_async.h>
-#include <dsd-neo/ui/ui_cmd.h>
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
