+++ tests/engine/test_engine_cleanup_audio.c
@@ -29,17 +29,6 @@ expect_true(const char* tag, int cond) {
     return 0;
 }
 
-int
-ui_start(dsd_opts* opts, dsd_state* state) { // NOLINT(misc-use-internal-linkage)
-    (void)opts;
-    (void)state;
-    return 0;
-}
-
-void
-ui_stop(void) { // NOLINT(misc-use-internal-linkage)
-}
-
 static int
 init_test_runtime(dsd_opts** opts_out, dsd_state** state_out) {
     // dsd_state is multi-megabyte; keep it off the function stack.
