+++ tests/ui/test_ui_history_state.c
@@ -6,7 +6,7 @@
 #include <assert.h>
 #include <stdio.h>
 
-#include <dsd-neo/ui/ui_history.h>
+#include <dsd-neo/app_control/history.h>
 
 int
 main(void) {
