+++ tests/ui/test_ui_tcp_audio_connect.c
@@ -4,13 +4,13 @@
  */
 
 #include <assert.h>
+#include <dsd-neo/app_control/services.h>
 #include <dsd-neo/core/audio.h>
 #include <dsd-neo/core/opts.h>
 #include <dsd-neo/dsp/resampler.h>
 #include <dsd-neo/io/tcp_input.h>
 #include <dsd-neo/platform/sockets.h>
 #include <dsd-neo/runtime/log.h>
-#include <dsd-neo/ui/menu_services.h>
 #include <string.h>
 #include "dsd-neo/core/opts_fwd.h"
 #include "dsd-neo/core/safe_api.h"
